+++ eudemo/eudemo/neutronics/run.py
@@ -13,6 +13,7 @@ from bluemira.neutronics.blanket_data import (
     create_materials,
     get_preset_physical_properties,
 )
+from bluemira.neutronics.geometry import TokamakDimensions
 from bluemira.neutronics.neutronics_axisymmetric import (
     NeutronicsReactor,
     NeutronicsReactorParameterFrame,
@@ -21,7 +22,6 @@ from bluemira.neutronics.openmc.solver import (
     OpenMCNeutronicsSolver,
 )
 from bluemira.neutronics.openmc.sources import make_pps_source
-from bluemira.neutronics.params import TokamakDimensions
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -32,7 +32,7 @@ if TYPE_CHECKING:
     from bluemira.base.parameter_frame import ParameterFrame
     from bluemira.base.reactor import ComponentManager
     from bluemira.geometry.wire import BluemiraWire
-    from bluemira.neutronics.params import PlasmaSourceParameters
+    from bluemira.neutronics.openmc.params import PlasmaSourceParameters
     from eudemo.blanket import Blanket
     from eudemo.ivc import IVCShapes
     from eudemo.vacuum_vessel import VacuumVessel
